django-poc/pyproject.toml
[tool.poetry]
name = "barc"
version = "0.1.0"
description = ""
authors = ["Glayson Olivieri"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
psycopg = "^3.2.1"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
django-extensions = "^3.2.3"
bpython = "^0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
